ROC-761-Ack-Malformed-Msg
This commit adds a field name check in the unmarshaller to reject unknown field names and return error
Additional house keeping - removed unused code
